requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
psycopg2-binary==2.9.9
google-generativeai==0.8.3
groq==0.11.0
python-dotenv==1.0.1
